+++ src/Parser/Parsing/Impl/AbstractDirectiveStreamProxy.cpp
@@ -1,11 +1,11 @@
 #include "AbstractDirectiveStreamProxy.h"
 
-TokenPos AbstractDirectiveStreamProxy::CreatePos(const ParserLine& line, const unsigned position)
+TokenPos AbstractDirectiveStreamProxy::CreatePos(const ParserLine& line, const size_t position)
 {
     return TokenPos(*line.m_filename, line.m_line_number, static_cast<int>(position + 1));
 }
 
-bool AbstractDirectiveStreamProxy::SkipWhitespace(const ParserLine& line, unsigned& position)
+bool AbstractDirectiveStreamProxy::SkipWhitespace(const ParserLine& line, size_t& position)
 {
     while (true)
     {
@@ -21,7 +21,7 @@ bool AbstractDirectiveStreamProxy::SkipWhitespace(const ParserLine& line, unsign
     return true;
 }
 
-bool AbstractDirectiveStreamProxy::ExtractInteger(const ParserLine& line, unsigned& position, int& value)
+bool AbstractDirectiveStreamProxy::ExtractInteger(const ParserLine& line, size_t& position, int& value)
 {
     if (position >= line.m_line.size())
         return false;
@@ -40,7 +40,7 @@ bool AbstractDirectiveStreamProxy::ExtractInteger(const ParserLine& line, unsign
     return false;
 }
 
-bool AbstractDirectiveStreamProxy::ExtractIdentifier(const ParserLine& line, unsigned& position)
+bool AbstractDirectiveStreamProxy::ExtractIdentifier(const ParserLine& line, size_t& position)
 {
     auto firstChar = true;
     while (true)
@@ -60,7 +60,7 @@ bool AbstractDirectiveStreamProxy::ExtractIdentifier(const ParserLine& line, uns
     }
 }
 
-bool AbstractDirectiveStreamProxy::MatchCharacter(const ParserLine& line, unsigned& position, char c)
+bool AbstractDirectiveStreamProxy::MatchCharacter(const ParserLine& line, size_t& position, char c)
 {
     if (position < line.m_line.size() && line.m_line[position] == c)
     {
@@ -71,12 +71,12 @@ bool AbstractDirectiveStreamProxy::MatchCharacter(const ParserLine& line, unsign
     return false;
 }
 
-bool AbstractDirectiveStreamProxy::MatchNextCharacter(const ParserLine& line, unsigned& position, char c)
+bool AbstractDirectiveStreamProxy::MatchNextCharacter(const ParserLine& line, size_t& position, char c)
 {
     return SkipWhitespace(line, position) && MatchCharacter(line, position, c);
 }
 
-bool AbstractDirectiveStreamProxy::MatchString(const ParserLine& line, unsigned& position, const char* str, unsigned len)
+bool AbstractDirectiveStreamProxy::MatchString(const ParserLine& line, size_t& position, const char* str, size_t len)
 {
     if (line.m_line.compare(position, len, str) == 0)
     {
@@ -87,12 +87,12 @@ bool AbstractDirectiveStreamProxy::MatchString(const ParserLine& line, unsigned&
     return false;
 }
 
-bool AbstractDirectiveStreamProxy::MatchNextString(const ParserLine& line, unsigned& position, const char* str, unsigned len)
+bool AbstractDirectiveStreamProxy::MatchNextString(const ParserLine& line, size_t& position, const char* str, size_t len)
 {
     return SkipWhitespace(line, position) && MatchString(line, position, str, len);
 }
 
-bool AbstractDirectiveStreamProxy::FindDirective(const ParserLine& line, unsigned& directiveStartPosition, unsigned& directiveEndPos)
+bool AbstractDirectiveStreamProxy::FindDirective(const ParserLine& line, size_t& directiveStartPosition, size_t& directiveEndPos)
 {
     directiveStartPosition = 0;
     for (; directiveStartPosition < line.m_line.size(); directiveStartPosition++)
